requirements.txt
click~=8.1.3
pandas~=1.4.3
scikit-learn~=1.1.1
fastapi~=0.78.0
pydantic~=1.9.1
numpy~=1.23.0
uvicorn
google-cloud-storage
google-cloud-aiplatform
kfp